add useActiveBalances hook
